Fold branch/commit into the build-info card; split it into its own file
Move the Branch and Commit rows out of the About section and into the
build-info card as a compact source strip below a divider (hidden on local
builds that weren't git-stamped). About now holds just the License link.

Extract the card and its AppInfo build-details extension into a dedicated
BuildInfoCard.swift so SettingsView stays under the file-length limit.
